+++ src/memory/CRAM.h
@@ -26,7 +26,7 @@ public:
   uint8_t   read(uint16_t);
   void      write(uint16_t, uint8_t);
   void      step(Bus_obj*){}
-  uint16_t  read_color_palette(uint8_t, uint8_t, uint8_t);
+  uint32_t  read_color_palette(uint8_t, uint8_t, uint8_t);
             ~CRAM(){}
 };
